+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.2
+
+* Fix file download functionality on Android
+* Update README with complete Android permission requirements
+
 ## 1.0.1
 
 * Update Readme file